Sky box depth fix and shadow mapping begun
messed around and managed to make it so that the skybox does not block
other objects, to do this though I had to mess with the way the visible
game objects are sorted as the skybox has to be the first object
rendered, other than that I have started shadow mapping. the depth
texture from the lights pov should be rendering properly, now I need to
generate them.